@@ -1,10 +1,13 @@
 // package memutils provides memory management utilities
 package memutils
 
-import "gorgonia.org/tensor"
+import (
+	"gorgonia.org/dtype"
+	"gorgonia.org/tensor"
+)
 
 // MemSize returns the size of the memory required
-func MemSize(dt tensor.Dtype, s tensor.Shape) int64 {
+func MemSize(dt dtype.Dtype, s tensor.Shape) int64 {
 	var elemSize int64
 	if s.IsScalar() {
 		elemSize = 1
